Move traildb builder into a separate package
