+++ .gitignore
@@ -1,3 +1,4 @@
+.idea
 target
 src/ffi/mod.rs
 *.tdb
