--- .gitmodules
@@ -1,3 +1,0 @@
-[submodule "traildb"]
-	path = traildb
-	url = https://github.com/traildb/traildb.git
