+++ Cargo.toml
@@ -2,7 +2,7 @@
 authors = ["Jay Kickliter <jay@kickliter.com>", "Kenan Sulayman <kenan@sly.mn>", "Knut Nesheim <knutin@gmail.com>"]
 build = "build.rs"
 name = "traildb"
-version = "0.4.0"
+version = "0.5.0"
 
 description = "Binding for TrailDB, an efficient tool for storing and querying series of events."
 license = "MIT"
@@ -12,13 +12,12 @@ repository = "https://github.com/traildb/traildb-rust.git"
 
 [build-dependencies]
 bindgen = "^0.20.2"
-pkg-config = "0.3.16"
-cc = { version = "1.0", features = ["parallel"] }
-glob = "0.3"
+pkg-config = "0.3"
 
 [dependencies]
 clang-sys = "0.12.0"
 libc = "0.2.21"
+traildb-sys = { version = "0.6.0", path = "./traildb-sys" }
 
 [dev-dependencies]
 prettytable-rs = "0.6.6"
