+++ traildb-sys/.gitignore
@@ -1,0 +1,2 @@
+/Cargo.lock
+/traildb-0.6/
