+++ traildb-sys/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "traildb-sys"
+authors = ["Tomas Sedlak <tomas.sedlak@simplestaking.com>"]
+description = "Builds traildb library"
+license = "MIT"
+version = "0.6.0"
+edition = "2018"
+
+[build-dependencies]
+cc = { version = "1.0", features = ["parallel"] }
+glob = "0.3"
+libflate = "0.1"
+pkg-config = "0.3"
+tar = "0.4"
