+++ traildb-sys/README.md
@@ -1,0 +1,5 @@
+# traildb builder
+
+This sub-crate is used to guild traildb from source code.
+
+Database source code is distributed in a file `traildb-0.6.tar.gz`.
