+++ traildb-sys/src/lib.rs
@@ -1,0 +1,1 @@
+// nothing to do here
